fix(players): prevent /players/null links for players without psdId
Make PlayerVM.href optional so players without a psdId don't produce
broken "/players/null" links. Filter non-routable players in team pages,
board pages, search, and static param generation. Fix OG image fallback
for null player lookups. Extract SanityPlayerBase to shared sanity/types
module to remove circular dependency.

Closes #882
