+++ apps/web/src/app/(main)/club/createBoardPage.tsx
@@ -11,7 +11,10 @@ import { notFound } from "next/navigation";
 import { runPromise } from "@/lib/effect/runtime";
 import { SanityService } from "@/lib/effect/services/SanityService";
 import { BestuurPage } from "@/components/club/BestuurPage/BestuurPage";
-import { toPlayerVM } from "@/lib/repositories/player.repository";
+import {
+  toPlayerVM,
+  type RoutablePlayerVM,
+} from "@/lib/repositories/player.repository";
 import {
   transformSanityStaffToMember,
   getSanityTeamTagline,
@@ -108,7 +111,11 @@ export function createBoardPage({
           tagline: getSanityTeamTagline(team),
           teamType: "club",
         }}
-        players={(team.players ?? []).map(toPlayerVM)}
+        players={
+          (team.players ?? [])
+            .filter((p) => p.psdId)
+            .map(toPlayerVM) as RoutablePlayerVM[]
+        }
         staff={(team.staff ?? []).map(transformSanityStaffToMember)}
       />
     );
